Update Window_BattleLog.d.ts
It's kinda wild, but works like a charm :) 

The only drawback is that it also accepts variable names on the `methodName` parameter, the best I could do to work around this was to return `never` in those cases, which should be enough. It's better than just a `string` anyway, and autocomplete works well with the `args` param too.